+++ src/main/java/com/programmers/smrtstore/domain/auth/jwt/JwtAuthentication.java
@@ -1,4 +1,4 @@
-package com.programmers.smrtstore.core.properties;
+package com.programmers.smrtstore.domain.auth.jwt;
 
 import java.util.Date;
 import lombok.Builder;
