+++ src/main/java/com/programmers/smrtstore/domain/auth/jwt/JwtAuthenticationToken.java
@@ -1,4 +1,4 @@
-package com.programmers.smrtstore.core.properties;
+package com.programmers.smrtstore.domain.auth.jwt;
 
 import java.util.Collection;
 import java.util.Objects;
